+++ registry/storage/registry.go
@@ -29,7 +29,7 @@ type registry struct {
 	db                           *datastore.DB
 }
 
-// manifestURLs holds regular expressions for controlling manifest URL whitelisting
+// manifestURLs holds regular expressions for controlling manifest URL allowlisting
 type manifestURLs struct {
 	allow *regexp.Regexp
 	deny  *regexp.Regexp